Fixing maven publishing
