+++ project/plugins.sbt
@@ -22,8 +22,9 @@ addSbtPlugin("org.scoverage" %% "sbt-coveralls" % "1.1.0")
 
 addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.1.10")
 
+addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
+
 if (sys.env.contains("MAVEN_PUBLISH")) {
-  addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
   addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "1.1")
 } else {
   addSbtPlugin("me.lessis" % "bintray-sbt" % "0.3.0")
